Struct deepl_api::DeepL

pub struct DeepL { /* private fields */ }

The main API entry point representing a DeepL developer account with an associated API key.

Example

See Example.

Error Handling

None of the functions will panic. Instead, the API methods usually return a Result which may contain an Error of one of the defined ErrorKinds with more information about what went wrong.

If you get an AuthorizationError, then something was wrong with your API key, for example.

Implementations

impl DeepL

Implements the actual REST API. See also the online documentation.

pub fn new(api_key: String) -> DeepL

Use this to create a new DeepL API client instance where multiple function calls can be performed. A valid api_key is required.

Should you ever need to use more than one DeepL account in our program, then you can create one instance for each account / API key.

pub fn usage_information(&self) -> Result<UsageInformation>

Retrieve information about API usage & limits. This can also be used to verify an API key without consuming translation contingent.

See also the vendor documentation.

pub fn source_languages(&self) -> Result<LanguageList>

Retrieve all currently available source languages.

See also the vendor documentation.

pub fn target_languages(&self) -> Result<LanguageList>

Retrieve all currently available target languages.

See also the vendor documentation.

pub fn translate(
&self,
options: Option<TranslationOptions>,
text_list: TranslatableTextList
) -> Result<Vec<TranslatedText>>

Translate one or more text chunks at once. You can pass in optional translation flags if you need non-default behaviour.

Please see the parameter documentation and the vendor documentation for details.

pub fn create_glossary(
&self,
name: String,
source_lang: String,
target_lang: String,
entries: String,
entries_format: GlossaryEntriesFormat
) -> Result<Glossary>

Create a glossary.

Please take a look at the vendor documentation for details.

pub fn list_glossaries(&self) -> Result<GlossaryListing>

List all glossaries.

Please take a look at the vendor documentation for details.

pub fn delete_glossary(&self, glossary_id: String) -> Result<Response>

Delete a glossary.

Please take a look at the vendor documentation for details.

pub fn get_glossary(&self, glossary_id: String) -> Result<Glossary>

Retrieve Glossary Details.

Please take a look at the vendor documentation for details.